+++ packages/flutter_shaders/example/lib/main.dart
@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -8,14 +8,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter_shaders/flutter_shaders.dart';
 
 void main() async {
-  final ui.FragmentProgram program =
-      await ui.FragmentProgram.fromAsset('shaders/inkwell.frag');
+  final ui.FragmentProgram program = await ui.FragmentProgram.fromAsset('shaders/inkwell.frag');
   runApp(MyApp(program: program));
 }
 
+/// The main application widget for the shader demo.
 class MyApp extends StatelessWidget {
+  /// Creates the [MyApp] demo widget.
   const MyApp({super.key, required this.program});
 
+  /// The fragment program used to power shader effects.
   final ui.FragmentProgram program;
 
   @override
@@ -23,30 +25,36 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: ThemeData(
-          primarySwatch: Colors.blue,
-          splashFactory: ShaderInkFeatureFactory(program, (
-            shader, {
-            required double animation,
-            required Color color,
-            required Offset position,
-            required Size referenceBoxSize,
-            required double targetRadius,
-            required TextDirection textDirection,
-          }) {
-            shader.setFloatUniforms((uniforms) => uniforms
+        primarySwatch: Colors.blue,
+        splashFactory: ShaderInkFeatureFactory(program, (
+          shader, {
+          required double animation,
+          required Color color,
+          required Offset position,
+          required Size referenceBoxSize,
+          required double targetRadius,
+          required TextDirection textDirection,
+        }) {
+          shader.setFloatUniforms(
+            (uniforms) => uniforms
               ..setFloat(animation)
               ..setColor(color, premultiply: true)
               ..setFloat(targetRadius)
-              ..setOffset(position));
-          })),
+              ..setOffset(position),
+          );
+        }),
+      ),
       home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
 
+/// The home page widget displaying counter and shader effects.
 class MyHomePage extends StatefulWidget {
+  /// Creates the home page with a given [title].
   const MyHomePage({super.key, required this.title});
 
+  /// The title displayed in the AppBar.
   final String title;
 
   @override
@@ -65,20 +73,13 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
+      appBar: AppBar(title: Text(widget.title)),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            const Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
+            const Text('You have pushed the button this many times:'),
+            Text('$_counter', style: Theme.of(context).textTheme.headlineMedium),
           ],
         ),
       ),