+++ packages/flutter_shaders/test/animated_sampler_test.dart
@@ -1,67 +1,43 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
+
 import 'package:flutter/material.dart';
 import 'package:flutter_shaders/flutter_shaders.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   testWidgets('does not call builder when not enabled', (tester) async {
-    final AnimatedSamplerBuilder builder =
-        expectAsync4((image, size, offset, canvas) {}, count: 0);
+    final AnimatedSamplerBuilder builder = expectAsync3((image, size, canvas) {}, count: 0);
 
     await tester.pumpWidget(
-      MaterialApp(
-        home: AnimatedSampler(
-          builder,
-          enabled: false,
-          child: SizedBox(),
-        ),
-      ),
+      MaterialApp(home: AnimatedSampler(builder, enabled: false, child: const SizedBox())),
     );
   });
 
   testWidgets('starts calling builder once enabled', (tester) async {
-    final AnimatedSamplerBuilder builder =
-        expectAsync4((image, size, offset, canvas) {}, count: 1);
+    final AnimatedSamplerBuilder builder = expectAsync3((image, size, canvas) {});
 
     await tester.pumpWidget(
-      MaterialApp(
-        home: AnimatedSampler(
-          builder,
-          enabled: false,
-          child: SizedBox(),
-        ),
-      ),
+      MaterialApp(home: AnimatedSampler(builder, enabled: false, child: const SizedBox())),
     );
 
-    await tester.pumpWidget(
-      MaterialApp(
-        home: AnimatedSampler(
-          builder,
-          enabled: true,
-          child: SizedBox(),
-        ),
-      ),
-    );
+    await tester.pumpWidget(MaterialApp(home: AnimatedSampler(builder, child: const SizedBox())));
   });
 
   testWidgets('rebuilds when child layer is updated', (tester) async {
-    final AnimatedSamplerBuilder builder =
-        expectAsync4((image, size, offset, canvas) {}, count: 2);
+    final AnimatedSamplerBuilder builder = expectAsync3((image, size, canvas) {}, count: 2);
 
     await tester.pumpWidget(
       MaterialApp(
         home: AnimatedSampler(
           builder,
-          child: RepaintBoundary(
-            child: CircularProgressIndicator(),
-          ),
+          child: const RepaintBoundary(child: CircularProgressIndicator()),
         ),
       ),
     );
 
     // Pump the next frame to animate `CircularProgressIndicator`.
-    await tester.pump(Duration(seconds: 1));
+    await tester.pump(const Duration(seconds: 1));
   });
 }