+++ packages/flutter_shaders/test/set_uniforms_test.dart
@@ -1,128 +1,82 @@
+// Copyright 2013 The Flutter Authors
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'dart:ui';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter_shaders/flutter_shaders.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:vector_math/vector_math.dart';
 
-class _MockFragmentShader implements FragmentShader {
-  final floats = <int, double>{};
-  final images = <int, Image>{};
+class _TestUniformsSetter extends FloatUniformsSetter {
+  _TestUniformsSetter();
+
+  int _index = 0;
+  final Map<int, double> floats = <int, double>{};
 
   @override
-  noSuchMethod(Invocation invocation) {
-    throw UnimplementedError();
-  }
-
-  @override
-  void setFloat(int index, double value) {
-    floats[index] = value;
+  void setFloat(double value) {
+    floats[_index++] = value;
   }
 }
 
 void main() {
   group('SetUniforms', () {
     test('setFloat', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setFloat(1.0);
-      });
+      setter.setFloat(1.0);
 
-      expect(shader.floats, {
-        0: 1.0,
-      });
+      expect(setter.floats, <int, double>{0: 1.0});
     });
 
     test('setFloats', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setFloats([1.0, 2.0, 3.0]);
-      });
+      setter.setFloats(<double>[1.0, 2.0, 3.0]);
 
-      expect(shader.floats, {
-        0: 1.0,
-        1: 2.0,
-        2: 3.0,
-      });
+      expect(setter.floats, <int, double>{0: 1.0, 1: 2.0, 2: 3.0});
     });
 
     test('setSize', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setSize(const Size(1.0, 2.0));
-      });
+      setter.setSize(const Size(1.0, 2.0));
 
-      expect(shader.floats, {
-        0: 1.0,
-        1: 2.0,
-      });
+      expect(setter.floats, <int, double>{0: 1.0, 1: 2.0});
     });
 
     test('setSizes', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setSizes(const [
-          Size(1.0, 2.0),
-          Size(3.0, 4.0),
-          Size(5.0, 6.0),
-        ]);
-      });
+      setter.setSizes(const <Size>[Size(1.0, 2.0), Size(3.0, 4.0), Size(5.0, 6.0)]);
 
-      expect(shader.floats, {
-        0: 1.0,
-        1: 2.0,
-        2: 3.0,
-        3: 4.0,
-        4: 5.0,
-        5: 6.0,
-      });
+      expect(setter.floats, <int, double>{0: 1.0, 1: 2.0, 2: 3.0, 3: 4.0, 4: 5.0, 5: 6.0});
     });
 
     test('setColor', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setColor(const Color(0xFF006600));
-      });
+      setter.setColor(const Color(0xFF006600));
 
-      expect(shader.floats, {
-        0: 0.0,
-        1: 0.4,
-        2: 0.0,
-        3: 1.0,
-      });
+      expect(setter.floats, <int, double>{0: 0.0, 1: 0.4, 2: 0.0, 3: 1.0});
     });
 
     test('setColor w/ premultiply', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setColor(const Color(0x00006600), premultiply: true);
-      });
+      setter.setColor(const Color(0x00006600), premultiply: true);
 
-      expect(shader.floats, {
-        0: 0.0,
-        1: 0.0,
-        2: 0.0,
-        3: 0.0,
-      });
+      expect(setter.floats, <int, double>{0: 0.0, 1: 0.0, 2: 0.0, 3: 0.0});
     });
 
     test('setColors', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setColors(const [
-          Color(0xFF006600),
-          Color(0xFF660000),
-          Color(0x66000066),
-        ]);
-      });
+      setter.setColors(const <Color>[Color(0xFF006600), Color(0xFF660000), Color(0x66000066)]);
 
-      expect(shader.floats, {
+      expect(setter.floats, <int, double>{
         0: 0.0,
         1: 0.4,
         2: 0.0,
@@ -134,25 +88,20 @@ void main() {
         8: 0.0,
         9: 0.0,
         10: 0.4,
-        11: 0.4
+        11: 0.4,
       });
     });
 
     test('setColors w/ premultiply', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setColors(
-          premultiply: true,
-          const [
-            Color(0xFF006600),
-            Color(0xFF660000),
-            Color(0x00000066),
-          ],
-        );
-      });
+      setter.setColors(premultiply: true, const <Color>[
+        Color(0xFF006600),
+        Color(0xFF660000),
+        Color(0x00000066),
+      ]);
 
-      expect(shader.floats, {
+      expect(setter.floats, <int, double>{
         0: 0.0,
         1: 0.4,
         2: 0.0,
@@ -164,54 +113,34 @@ void main() {
         8: 0.0,
         9: 0.0,
         10: 0.0,
-        11: 0.4
+        11: 0.0,
       });
     });
 
     test('setOffset', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setOffset(const Offset(1.0, 2.0));
-      });
+      setter.setOffset(const Offset(1.0, 2.0));
 
-      expect(shader.floats, {
-        0: 1.0,
-        1: 2.0,
-      });
+      expect(setter.floats, <int, double>{0: 1.0, 1: 2.0});
     });
 
     test('setOffsets', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setOffsets(const [
-          Offset(1.0, 2.0),
-          Offset(3.0, 4.0),
-          Offset(5.0, 6.0),
-        ]);
-      });
+      setter.setOffsets(const <Offset>[Offset(1.0, 2.0), Offset(3.0, 4.0), Offset(5.0, 6.0)]);
 
-      expect(shader.floats, {
-        0: 1.0,
-        1: 2.0,
-        2: 3.0,
-        3: 4.0,
-        4: 5.0,
-        5: 6.0,
-      });
+      expect(setter.floats, <int, double>{0: 1.0, 1: 2.0, 2: 3.0, 3: 4.0, 4: 5.0, 5: 6.0});
     });
 
     test('setVector', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setVector(Vector2(1.0, 2.0));
-        setter.setVector(Vector3(3.0, 4.0, 5.0));
-        setter.setVector(Vector4(6.0, 7.0, 8.0, 9.0));
-      });
+      setter.setVector(Vector2(1.0, 2.0));
+      setter.setVector(Vector3(3.0, 4.0, 5.0));
+      setter.setVector(Vector4(6.0, 7.0, 8.0, 9.0));
 
-      expect(shader.floats, {
+      expect(setter.floats, <int, double>{
         0: 1.0,
         1: 2.0,
         2: 3.0,
@@ -225,32 +154,23 @@ void main() {
     });
 
     test('setMatrix2', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setMatrix2(Matrix2(1.0, 2.0, 3.0, 4.0));
-      });
+      setter.setMatrix2(Matrix2(1.0, 2.0, 3.0, 4.0));
 
-      expect(shader.floats, {
-        0: 1.0,
-        1: 2.0,
-        2: 3.0,
-        3: 4.0,
-      });
+      expect(setter.floats, <int, double>{0: 1.0, 1: 2.0, 2: 3.0, 3: 4.0});
     });
 
     test('setMatrix2s', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setMatrix2s([
-          Matrix2(1.0, 2.0, 3.0, 4.0),
-          Matrix2(5.0, 6.0, 7.0, 8.0),
-          Matrix2(9.0, 10.0, 11.0, 12.0),
-        ]);
-      });
+      setter.setMatrix2s(<Matrix2>[
+        Matrix2(1.0, 2.0, 3.0, 4.0),
+        Matrix2(5.0, 6.0, 7.0, 8.0),
+        Matrix2(9.0, 10.0, 11.0, 12.0),
+      ]);
 
-      expect(shader.floats, {
+      expect(setter.floats, <int, double>{
         0: 1.0,
         1: 2.0,
         2: 3.0,
@@ -267,13 +187,11 @@ void main() {
     });
 
     test('setMatrix3', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setMatrix3(Matrix3(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0));
-      });
+      setter.setMatrix3(Matrix3(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0));
 
-      expect(shader.floats, {
+      expect(setter.floats, <int, double>{
         0: 1.0,
         1: 2.0,
         2: 3.0,
@@ -287,17 +205,15 @@ void main() {
     });
 
     test('setMatrix3s', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setMatrix3s([
-          Matrix3(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0),
-          Matrix3(10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0, 17.0, 18.0),
-          Matrix3(19.0, 20.0, 21.0, 22.0, 23.0, 24.0, 25.0, 26.0, 27.0),
-        ]);
-      });
+      setter.setMatrix3s(<Matrix3>[
+        Matrix3(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0),
+        Matrix3(10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0, 17.0, 18.0),
+        Matrix3(19.0, 20.0, 21.0, 22.0, 23.0, 24.0, 25.0, 26.0, 27.0),
+      ]);
 
-      expect(shader.floats, {
+      expect(setter.floats, <int, double>{
         0: 1.0,
         1: 2.0,
         2: 3.0,
@@ -329,10 +245,10 @@ void main() {
     });
 
     test('setMatrix4', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setMatrix4(Matrix4(
+      setter.setMatrix4(
+        Matrix4(
           1.0,
           2.0,
           3.0,
@@ -349,10 +265,10 @@ void main() {
           14.0,
           15.0,
           16.0,
-        ));
-      });
+        ),
+      );
 
-      expect(shader.floats, {
+      expect(setter.floats, <int, double>{
         0: 1.0,
         1: 2.0,
         2: 3.0,
@@ -373,50 +289,48 @@ void main() {
     });
 
     test('setMatrix4s', () {
-      final shader = _MockFragmentShader();
+      final setter = _TestUniformsSetter();
 
-      shader.setFloatUniforms((setter) {
-        setter.setMatrix4s([
-          Matrix4(
-            1.0,
-            2.0,
-            3.0,
-            4.0,
-            5.0,
-            6.0,
-            7.0,
-            8.0,
-            9.0,
-            10.0,
-            11.0,
-            12.0,
-            13.0,
-            14.0,
-            15.0,
-            16.0,
-          ),
-          Matrix4(
-            17.0,
-            18.0,
-            19.0,
-            20.0,
-            21.0,
-            22.0,
-            23.0,
-            24.0,
-            25.0,
-            26.0,
-            27.0,
-            28.0,
-            29.0,
-            30.0,
-            31.0,
-            32.0,
-          ),
-        ]);
-      });
+      setter.setMatrix4s(<Matrix4>[
+        Matrix4(
+          1.0,
+          2.0,
+          3.0,
+          4.0,
+          5.0,
+          6.0,
+          7.0,
+          8.0,
+          9.0,
+          10.0,
+          11.0,
+          12.0,
+          13.0,
+          14.0,
+          15.0,
+          16.0,
+        ),
+        Matrix4(
+          17.0,
+          18.0,
+          19.0,
+          20.0,
+          21.0,
+          22.0,
+          23.0,
+          24.0,
+          25.0,
+          26.0,
+          27.0,
+          28.0,
+          29.0,
+          30.0,
+          31.0,
+          32.0,
+        ),
+      ]);
 
-      expect(shader.floats, {
+      expect(setter.floats, <int, double>{
         0: 1.0,
         1: 2.0,
         2: 3.0,
@@ -451,5 +365,19 @@ void main() {
         31: 32.0,
       });
     });
+
+    test('UniformsSetter with FragmentShader', () async {
+      TestWidgetsFlutterBinding.ensureInitialized();
+      final FragmentProgram program = await FragmentProgram.fromAsset('shaders/ink_sparkle.frag');
+      final FragmentShader shader = program.fragmentShader();
+
+      final int nextIndex = shader.setFloatUniforms((UniformsSetter setter) {
+        expect(setter.shader, shader);
+        setter.setFloat(1.0);
+        setter.setFloat(2.0);
+      });
+
+      expect(nextIndex, 2);
+    }, skip: kIsWeb); // The ink_sparkle asset is only available in VM test harnesses.
   });
 }